requirements.txt
streamlit
pydrive
pymupdf
python-docx
pandas
openai
